Metrics API: added 'meta' and 'help'

+++ adsabs/modules/bibutils/metricsmodels/models.py
@@ -79,17 +79,17 @@ class Statistics():
         # get number of refereed entries
         cls.number_of_refereed_entries = len(refereed_values)
         # get normalized value
-        cls.normalized_value = '%.1f' % round(vector_product(values,weights), 1)
+        cls.normalized_value = float('%.1f' % round(vector_product(values,weights), 1))
         # get refereed normalized value
-        cls.refereed_normalized_value = '%.1f' % round(vector_product(refereed_values,refereed_weights), 1)
+        cls.refereed_normalized_value = float('%.1f' % round(vector_product(refereed_values,refereed_weights), 1))
         # get mean value of values
-        cls.mean_value = '%.1f' % round(mean(values), 1)
+        cls.mean_value = float('%.1f' % round(mean(values), 1))
         # get mean value of refereed values
-        cls.refereed_mean_value = '%.1f' % round(mean(refereed_values), 1)
+        cls.refereed_mean_value = float('%.1f' % round(mean(refereed_values), 1))
         # get median value of values
-        cls.median_value = '%.1f' % round(median(values), 1)
+        cls.median_value = float('%.1f' % round(median(values), 1))
         # get median value of refereed values
-        cls.refereed_median_value = '%.1f' % round(median(refereed_values), 1)
+        cls.refereed_median_value = float('%.1f' % round(median(refereed_values), 1))
         # get total of values
         cls.total_value = sum(values)
         # get total of refereed values
@@ -160,12 +160,12 @@ class Metrics():
             riq = "NA"
         cls.h_index = h
         cls.g_index = g
-        cls.m_index = '%.1f' % round(float(h)/float(cls.time_span), 2)
+        cls.m_index = float('%.1f' % round(float(h)/float(cls.time_span), 2))
         cls.i10_index = len(filter(lambda a: a >= 10, citations))
         cls.i100_index= len(filter(lambda a: a >= 100, citations))
-        cls.e_index = '%.1f' % round(e,1)
-        cls.tori = '%.1f' % round(tori,1)
-        cls.riq  = '%.1f' % round(riq,1)
+        cls.e_index = float('%.1f' % round(e,1))
+        cls.tori = float('%.1f' % round(tori,1))
+        cls.riq  = float('%.1f' % round(riq,1))
         cls.read10 = int(round(read10))
 
         cls.post_process()
